habit_tracker_api_v1/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
autopep8==2.0.4
black==23.9.1
click==8.1.7
decorator==5.1.1
dj-database-url==1.2.0
Django==4.2.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
exceptiongroup==1.2.2
executing==2.1.0
gunicorn==21.2.0
importlib_resources==6.4.5
inflection==0.5.1
iniconfig==2.0.0
ipython==8.28.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
mysqlclient==2.2.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
Pillow==10.0.1
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.3.3
pytest-django==4.5.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
swagger-spec-validator==3.0.4
tomli==2.0.2
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
wcwidth==0.2.13
whitenoise==6.5.0
